Informatics Lunch Practical

Informatics Lunch Practical

Functions

Functions

Position.jl

ilp.Position.LongLat — Type

struct LongLat

A structure that stores position related property for an object

Fields

longitude: the longitude of the position of any point

latitude: the latitude of the position of any point

angle: the angle of the drone at a specific point, in integer. It is required to be a multiple of 10.

ilp.Position.Line2D — Type

Line2D(s::LongLat, e::LongLat)

Line2D represents a line by its starting and ending point

ilp.Position.isconfined — Method

isconfined(coord::LongLat)::Bool

Takes a coordinate and check whether it's in the confinement zone. returns true if it is in the zone.

ilp.Position.intersects — Method

intersects(e::Line2D, l::Line2D)::Bool

(a,b)→(c,d) intersects (p,q) → (r,s) Check whether two lines intersects

ilp.Position.distanceto — Method

distanceto(a::LongLat, b::LongLat)::Float64

Calculate the distance between two coordinates

ilp.Position.distanceto — Method

distanceto(a::LongLat, b::LongLat, noflyzone::Vector{Vector{Line2D}})::Float64

Get the distance between two coordinates, if it intersects the no-fly zone it will be set to a high value

ilp.Position.closeto — Method

closeto(coord::LongLat, point::LongLat)::Bool

check whether two points are close to each other

ilp.Position.nextPosition — Method

nextPosition(coord::LongLat, θ::Int64)::LongLat

Get the next position of the drone